[refactor] 비밀번호 검증, 성공합치기
