+++ src/main/java/dev/woori/wooriLearn/WooriLearnApplication.java
@@ -3,13 +3,17 @@ package dev.woori.wooriLearn;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 @SpringBootApplication
 @EnableJpaAuditing
 public class WooriLearnApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(WooriLearnApplication.class, args);
+        SpringApplication.run(WooriLearnApplication.class, args);
+
+
 	}
 
 }
