+++ src/main/java/dev/woori/wooriLearn/domain/edubankapi/entity/EducationalAccount.java
@@ -16,7 +16,6 @@ import java.util.List;
 
 @Entity
 @Getter
-@Setter
 @Table(name = "educational_account")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
@@ -37,7 +36,7 @@ public class EducationalAccount {
     private Integer balance;
 
     // 계좌 비밀번호
-    @Column(name = "account_password", nullable = false, length = 4)
+    @Column(name = "account_password", nullable = false, length = 100)
     private String accountPassword;
 
     // 계좌명
